fix: [debug] allow unit tests to assert on gd/cb_adf state
